Extract tag-related functions from app-state
