+++ lib/flux/app-state.js
@@ -2,14 +2,11 @@ import { get, partition, some } from 'lodash';
 import update from 'react-addons-update';
 import Debug from 'debug';
 import ActionMap from './action-map';
-import isEmailTag from '../utils/is-email-tag';
 import filterNotes from '../utils/filter-notes';
-import throttle from '../utils/throttle';
 import analytics from '../analytics';
 import { util as simperiumUtil } from 'simperium';
 
 const debug = Debug('appState');
-const typingThrottle = 3000;
 
 export const actionMap = new ActionMap({
   namespace: 'App',
@@ -174,97 +171,6 @@ export const actionMap = new ActionMap({
       return update(state, {
         editingTags: { $set: !state.editingTags },
       });
-    },
-
-    newTag: {
-      creator({ tagBucket, name }) {
-        return () => {
-          // tag.id must be the tag name in lower case and percent escaped
-          const tagId = encodeURIComponent(name.toLowerCase());
-          tagBucket.update(tagId, { name });
-        };
-      },
-    },
-
-    renameTag: {
-      creator({ tagBucket, noteBucket, tag, name }) {
-        return (dispatch, getState) => {
-          let oldTagName = tag.data.name;
-          if (oldTagName === name) {
-            return;
-          }
-
-          let { notes } = getState().appState;
-          let changedNoteIds = [];
-
-          tag.data.name = name;
-
-          // update the tag bucket but don't fire a sync immediately
-          tagBucket.update(tag.id, tag.data, { sync: false });
-
-          // similarly, update the note bucket
-          for (let i = 0; i < notes.length; i++) {
-            let note = notes[i];
-            let noteTags = note.data.tags || [];
-            let tagIndex = noteTags.indexOf(oldTagName);
-
-            if (tagIndex !== -1) {
-              noteTags.splice(tagIndex, 1, name);
-              note.data.tags = noteTags.filter(
-                noteTag => noteTag !== oldTagName
-              );
-              noteBucket.update(note.id, note.data, { sync: false });
-              changedNoteIds.push(note.id);
-            }
-          }
-
-          throttle(tag.id, typingThrottle, () => {
-            tagBucket.touch(tag.id, () => {
-              dispatch(this.action('loadTags', { tagBucket }));
-
-              for (let i = 0; i < changedNoteIds.length; i++) {
-                noteBucket.touch(changedNoteIds[i]);
-              }
-            });
-          });
-        };
-      },
-    },
-
-    trashTag: {
-      creator({ tagBucket, noteBucket, tag }) {
-        return (dispatch, getState) => {
-          var { notes } = getState().appState;
-          var tagName = tag.data.name;
-
-          for (let i = 0; i < notes.length; i++) {
-            let note = notes[i];
-            let noteTags = note.data.tags || [];
-            let newTags = noteTags.filter(noteTag => noteTag !== tagName);
-
-            if (newTags.length !== noteTags.length) {
-              note.data.tags = newTags;
-              noteBucket.update(note.id, note.data);
-            }
-          }
-
-          tagBucket.remove(tag.id, () => {
-            dispatch(this.action('loadTags', { tagBucket }));
-          });
-        };
-      },
-    },
-
-    reorderTags: {
-      creator({ tagBucket, tags }) {
-        return () => {
-          for (let i = 0; i < tags.length; i++) {
-            let tag = tags[i];
-            tag.data.index = i;
-            tagBucket.update(tag.id, tag.data);
-          }
-        };
-      },
     },
 
     search(state, { filter }) {
@@ -558,42 +464,6 @@ export const actionMap = new ActionMap({
       },
     },
 
-    updateNoteTags: {
-      creator({ noteBucket, tagBucket, note, tags }) {
-        return (dispatch, getState) => {
-          if (note) {
-            let state = getState().appState;
-
-            note.data.tags = tags;
-            note.data.modificationDate = Math.floor(Date.now() / 1000);
-            noteBucket.update(note.id, note.data);
-
-            dispatch(this.action('selectNote', { note }));
-
-            let currentTagNames = state.tags.map(tag => tag.data.name);
-            for (let i = 0; i < tags.length; i++) {
-              let tag = tags[i];
-
-              if (currentTagNames.indexOf(tag) !== -1) {
-                continue;
-              }
-
-              if (isEmailTag(tag)) {
-                continue;
-              }
-
-              dispatch(
-                this.action('newTag', {
-                  tagBucket,
-                  name: tag,
-                })
-              );
-            }
-          }
-        };
-      },
-    },
-
     trashNote: {
       creator({ noteBucket, note, previousIndex }) {
         return dispatch => {
@@ -679,29 +549,6 @@ export const actionMap = new ActionMap({
         showNavigation: { $set: false },
         editingTags: { $set: false },
       });
-    },
-
-    loadTags: {
-      creator({ tagBucket }) {
-        return (dispatch, getState) => {
-          const sortTagsAlpha = getState().settings.sortTagsAlpha;
-          tagBucket.query(db => {
-            var tags = [];
-            db
-              .transaction('tag')
-              .objectStore('tag')
-              .openCursor(null, 'prev').onsuccess = e => {
-              var cursor = e.target.result;
-              if (cursor) {
-                tags.push(cursor.value);
-                cursor.continue();
-              } else {
-                dispatch(this.action('tagsLoaded', { tags, sortTagsAlpha }));
-              }
-            };
-          });
-        };
-      },
     },
 
     tagsLoaded(state, { tags, sortTagsAlpha }) {
